shellia_controller.py
"""
🧠 SHELLIA CONTROLLER - IA Contrôleur de Maxis
Pilote le bot e-commerce Maxis depuis une VM séparée
"""

import discord
from discord.ext import commands, tasks
import aiohttp
import asyncio
from datetime import datetime, timedelta
from typing import Dict, Optional, Any, List
import json
import logging

logger = logging.getLogger(__name__)


class ShelliaController(commands.Bot):
    """
    🧠 Shellia - L'IA qui contrôle Maxis
    """
    
    def __init__(self):
        intents = discord.Intents.default()
        intents.message_content = True
        intents.members = True
        
        super().__init__(
            command_prefix='!',
            intents=intents,
            help_command=None
        )
        
        # Configuration Maxis
        self.maxis_config = {
            'api_url': 'http://maxis-vm:8080/api',
            'api_key': None,
            'vm_status': 'unknown',
            'last_ping': None
        }
        
        # État de Maxis
        self.maxis_status = {
            'online': False,
            'version': None,
            'uptime': None,
            'stats': {},
            'last_update': None
        }
        
    async def setup_hook(self):
        """Initialisation"""
        print('🧠 Shellia Controller initialisée')
        self.maxis_monitor.start()
        
    async def on_ready(self):
        """Shellia prête"""
        print(f'🧠 Shellia connectée: {self.user}')
        await self.change_presence(
            activity=discord.Activity(
                type=discord.ActivityType.watching,
                name="Maxis | !help"
            )
        )
        
    @tasks.loop(seconds=30)
    async def maxis_monitor(self):
        """Surveille Maxis"""
        try:
            status = await self._ping_maxis()
            self.maxis_status['online'] = status
            self.maxis_status['last_update'] = datetime.utcnow()
        except Exception as e:
            logger.error(f"Erreur surveillance: {e}")
            self.maxis_status['online'] = False
            
    async def _ping_maxis(self) -> bool:
        """Ping Maxis"""
        try:
            async with aiohttp.ClientSession() as session:
                async with session.get(
                    f"{self.maxis_config['api_url']}/health",
                    timeout=aiohttp.ClientTimeout(total=5)
                ) as resp:
                    return resp.status == 200
        except:
            return False
            
    async def _send_command_to_maxis(self, endpoint: str, data: Dict) -> Dict:
        """Envoie une commande à Maxis"""
        try:
            async with aiohttp.ClientSession() as session:
                async with session.post(
                    f"{self.maxis_config['api_url']}/{endpoint}",
                    json=data,
                    headers={'X-API-Key': self.maxis_config['api_key'] or ''},
                    timeout=aiohttp.ClientTimeout(total=10)
                ) as resp:
                    if resp.status == 200:
                        return await resp.json()
                    return {'success': False, 'error': f'Status {resp.status}'}
        except Exception as e:
            return {'success': False, 'error': str(e)}


# ... (commandes similaires au fichier précédent)

if __name__ == "__main__":
    import os
    token = os.getenv('SHELLIA_DISCORD_TOKEN')
    if not token:
        print("❌ SHELLIA_DISCORD_TOKEN manquant")
        exit(1)
    bot = ShelliaController()
    bot.run(token)
