requirements.txt
# Discord
discord.py>=2.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0

# Base de données
supabase>=2.0.0
psycopg2-binary>=2.9.0

# IA
requests>=2.31.0
google-generativeai>=0.3.0

# Sécurité
cryptography>=41.0.0
python-dotenv>=1.0.0

# Utilitaires
Pillow>=10.0.0
python-dateutil>=2.8.0
